+++ src/Layers/xrRender/dxRenderDeviceRender.cpp
@@ -189,7 +189,7 @@ void dxRenderDeviceRender::Create(SDL_Window* window, u32 &dwWidth, u32 &dwHeigh
 			return;
 
 		ImGui::SetNextWindowPos(ImVec2(0, 0));
-		ImGui::SetNextWindowSize(ImVec2(Device.TargetWidth, Device.TargetHeight));
+		ImGui::SetNextWindowSize(ImVec2((float)Device.TargetWidth, (float)Device.TargetHeight));
 		ImGui::SetNextWindowBgAlpha(0.0f);
 
 		ImGui::PushStyleVar(ImGuiStyleVar_WindowBorderSize, 0.0f);
